Fixed test that was failing with chrischrischris's help by moving to new file.
